feat: package json interface
